requirements.txt
# Core BigQuery and Google Cloud dependencies
google-cloud-bigquery>=3.11.0
google-cloud-storage>=2.10.0
google-cloud-aiplatform>=1.35.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Jupyter and visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# AI/ML specific
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Additional utilities
requests>=2.31.0
pillow>=10.0.0
opencv-python>=4.8.0
